@@ -1,9 +1,14 @@
-<?php declare(strict_types=1);
+<?php
+
+declare(strict_types=1);
 
 namespace WyriHaximus\Twig;
 
+use function function_exists;
+
 // @codeCoverageIgnoreStart
-if (!\function_exists('WyriHaximus\Twig\render')) {
+if (! function_exists('WyriHaximus\Twig\render')) {
     require __DIR__ . '/functions.php';
 }
+
 // @codeCoverageIgnoreEnd
